Support pdo odbc (#5053)
* init

* add

* tests

* disable pdo_odbc by default

* fix

* db-init

* fix 8.0

* add more info

* fix tests

* fix tests

* fix tests[3]

* fix tests[4]

* fix tests[5]

+++ src/core/base.cc
@@ -307,7 +307,7 @@ SW_API void swoole_set_log_file(const char *file) {
     }
 }
 
-SW_API void swoole_set_trace_flags(int flags) {
+SW_API void swoole_set_trace_flags(long flags) {
     SwooleG.trace_flags = flags;
 }
